Alerting through the server now works
The jest tests were removed since the structure of people-context has changed in such a way that makes it possible to have a two-way connection with the server through the WebSocket API and send alerts from one user to the other. The IP address used to connect to the server will have to change to be more consistent before this app becomes usable.